# Streamlined compilation