--- a/resources/examples/Showtimes/compiled/1.1.2/apiblueprint/representations/Coded error.apib
+++ b/resources/examples/Showtimes/compiled/1.1.2/apiblueprint/representations/Coded error.apib
@@ -1,3 +1,0 @@
-## Coded error
-- `error` (string) - User-friendly error message
-- `error_code` (number) - Error code