--- a/src/Application.php
+++ b/src/Application.php
@@ -18,4 +18,43 @@ class Application
      * @var string
      */
     const DOT_NOTATION_ANNOTATION_PARAMETER_TYPE_KEY = '__PARAMETER_TYPE__';
+
+    /** @var Container */
+    protected $container;
+
+    /**
+     * @param string $config_path
+     * @param bool $load_bootstrap
+     */
+    public function __construct(string $config_path, bool $load_bootstrap = true)
+    {
+        $this->container = new Container([
+            'config.path' => $config_path,
+            'config.load_bootstrap' => $load_bootstrap
+        ]);
+    }
+
+    /**
+     * @return Container
+     */
+    public function getContainer(): Container
+    {
+        return $this->container;
+    }
+
+    /**
+     * @param Container $container
+     */
+    public function setContainer(Container $container): void
+    {
+        $this->container = $container;
+    }
+
+    /**
+     * @return Config
+     */
+    public function getConfig(): Config
+    {
+        return $this->container->getConfig();
+    }
 }