--- a/src/Compiler/Changelog/Formats/Markdown.php
+++ b/src/Compiler/Changelog/Formats/Markdown.php
@@ -11,7 +11,7 @@ class Markdown extends Json
     /** @var string */
     protected $output_format = Changelog::FORMAT_MARKDOWN;
 
-    /** @var array */
+    /** @var string */
     protected $markdown = '';
 
     /**