This rewrites the compiler to resolve a security bug where representations that were only ever used within a private manner were exposed during public compilation.

The way the compiler now works is that it compiles documentation from the resource level down.

```
> resource
  |- representation
> resource
  |- representation
    |- representation
  |- representation
    |- representation
      |- representation
```

Previously the way it worked was:

```
> resource
> resource
> representation
> representation
```

## Benchmarks

| Command | Old | New | Improvement |
| :--- | :--- | :--- | :--- |
| ./vendor/bin/mill compile specs/ | 42.2041s | 54.9297 | -23.17% |
| ./vendor/bin/mill compile --format=apiblueprint specs/ | 8.4887s | 12.3761s  | -31.41% |
| ./vendor/bin/mill changelog specs/ | 2.374s | 3.9636s | -40.10% |
| ./vendor/bin/mill error specs/ | 2.044s | 3.8725s | -47.22% |

*Average execution time over 10 runs. Tests run against the Vimeo API codebase.*